1.20.1-2.4.1 - Tales, Trails, and Tools! (Github/Open Source Transition) - Fixed a bug in the 1.20 port causing the plant fiber not to drop from grass when using knives - Added an Energized Di-Emerald Knife - this was an oversight on my part, it was supposed to be in when knives released!    - Functions identically to all the other Energized tools, consumes FE and can be recharged via lightning strike - Added Copper Tools    - Slightly slower than Iron Tools, same mining level, significantly more durability
   - Added every vanilla tool variant as well as every Instrumentus tool variant

   - This is to add parity with the next feature...

- Added Multiple Tiers of Brushes

   - Wood, Stone, Iron, Gold, Diamond, Netherite, and Energized brushes!
   - They all brush at the same speed, but have varying durability, all (except wood) better than the original variant.
- Fixed [#12] Hammers do not AOE Harvest Properly
   - Hammers now harvest based on the targeted blockface.
- Added a AOE Outline renderer for the hammers (may add this for other AOE tools in the future)

+++ src/main/java/com/beanbot/instrumentus/common/items/PaxelItem.java
@@ -35,7 +35,7 @@ public class PaxelItem extends DiggerItem {
             SHOVEL_DIG, SHOVEL_FLATTEN,
             PICKAXE_DIG);
 
-    public PaxelItem(Tier material, int attackDamageIn, float attackSpeedIn, Item.Properties properties) {
+    public PaxelItem(Tier material, float attackDamageIn, float attackSpeedIn, Item.Properties properties) {
         super(attackDamageIn, attackSpeedIn, material, GeneratorBlockTags.MINEABLE_WITH_PAXEL, properties);
         this.material = material;
     }
